Move AnyDataReaderListener to separate module
