+++ dds/src/subscription/subscriber.rs
@@ -2,7 +2,9 @@ use crate::{
     domain::{
         domain_participant::DomainParticipant, domain_participant_factory::THE_PARTICIPANT_FACTORY,
     },
-    implementation::dds_impl::builtin_subscriber::BuiltInSubscriber,
+    implementation::dds_impl::{
+        any_data_reader_listener::AnyDataReaderListener, builtin_subscriber::BuiltInSubscriber,
+    },
     infrastructure::{
         condition::StatusCondition,
         error::{DdsError, DdsResult},
@@ -13,11 +15,7 @@ use crate::{
 };
 use crate::{
     implementation::{
-        dds_impl::{
-            user_defined_data_reader::AnyDataReaderListener,
-            user_defined_subscriber::UserDefinedSubscriber,
-        },
-        utils::shared_object::DdsWeak,
+        dds_impl::user_defined_subscriber::UserDefinedSubscriber, utils::shared_object::DdsWeak,
     },
     infrastructure::instance::InstanceHandle,
 };
